fix: linker scripts
Previously, we'd copied down the linker scripts locally, but that
makes it hard to track changes upstream. Instead, and by way of the
example at https://github.com/esp-rs/esp-template, depend on the
upstream linker scripts that the esp-rs hal puts in the linker search
paths (but cannot, on its own, cause us to use).