feat: handle backspace, emit character codes
Introduces a new Cursor that can be moved by a (positive or negative)
row & column offset, and uses that to interpret "delete" by moving the
cursor backwards and emitting a space character. NB: because we only
know our current position, we wrap back to the rightmost column of a
previous line when deleting when the cursor is in column zero.

There's also a (possible) implementation of handling newline and
carriage return characters, but those interact unexpectedly with how
screen et. al configure the serial terminal device (it only sends
'\r's without extra configuration, see .screenrc).

Finally, changes the default handling to print escaped characters, so
e.g. '\x08' would be rendered as '\u{08}' instead of simply moving the
cursor to the right. The hope is to make it easier to identify and
implement unhandled control sequences going forward.