feat: move pins out of gpio interrupts
In order to support a full duplex channel like PS/2, it's important to
be able to switch dynamically from listening to input on a pin (e.g. via
an interrupt) to an output and back again.

Additionally, an important safety concern is to ensure only a single
exclusive owner is able to read/write to the  pin at a time, so as to
avoid unintentionally triggering the GPIO interrupt with a write.

This change provides a mechanism to move a GPIO pin that's currently
awaiting an interrupt back into the caller's control so they may use it
as an output. The `keyboard` module uses this capability to implement
`send_reset` that asks a keyboard over PS/2 to reset its state (and
perform a self-test) before returning to awaiting device-driven
interrupts (like the self-test response).

NB: I haven't tested this code, only allowed the compiler to prove it
correct.

Co-Authored-By: Eric Douglass <dougli1sqrd@users.noreply.github.com>