fix: instead of panic, code now contains behavior
This sequence of characters causes the rust compiler to emit object code
that, when executed, produces the desired effect of being able to safely
take an interrupt pin back when holding an appropriate PinRef receipt.
The other sequence of characters did not do this, and so I like this
sequence of characters more.

There's something confusing to me going on with respect to the
information required to check whether the trait object really is a
PinInterrupt and dereference the data half of the thick pointer, but it
seems like the default implementation maybe hid that behind another
Box<dyn Any>? I guess it is easier to "see" how to build the right kind
of `dyn Any` out of a `Box<PinInterrupt>` than a `Box<Irq>`? But the
latter type there doesn't make sense to me, so there's at least one
thing about how traits with default methods get reified that I'm not
understanding.