refactor: interrupt_mask wrapper
This change introduces a new facility for interrupt handlers to raise
the priority threshold and re-enable interrupts so that handlers can
"nest" in a controlled way, extracted from the direct implementation in
UART0 prior.

Also, adds a few optimizations to the GPIO interrupt path to try and
address the "screen glitches when input is processed" issue. Sadly, they
do not appear to be sufficient at this time.

Finally, revs our hal fork to be (more) in sync with upstream to prepare
for more similar optimizations on that end to get us over the goal line.