refactor: treat main loop as event reactor
The main idea behind this change is to make the data coupling between
components both 1) explicit, and 2) in a single place. Previously,
`update` had to consume only a single key event at a time, regardless of
how many were produced, lest the input not "see" a key as it went by.

Now, that coupling is more apparent (especially if we view the body of
the main loop as "polling a series of devices with the latest-known
state"), and we've got a lot more control over the relative timing (i.e.
it's possible to move the `flush_and_parse` call somewhere after the
continue to ensure that we consume all of the key events before we parse
more).