fix: update hal, remove r0: fixes .rwtext init
Previously, under certain conditions, the compiler would emit a call to
`memcpy` to implement the portion of __pre_init that was responsible for
populating the .rwtext segment in RAM (loading it from flash). When
`memcpy` was moved into the .rwtext segment, hilarity ensued.

This change brings in, among other things, the upstream removal of r0,
a rust-level implementation of initialization routines, which appears to
have been deprecated in favor of explicit hand-written assembly
instructions in order to better handle cases just like this one.