perf: memcpy, memset now operate over words
Previously, the extremely simple implemenation copied one byte at a
time, but this appears to have had severe performance impacts (on the
order of a thousand cycles for one to two hundred bytes copied).

Instead, we perform just enough single byte copies to be word-sized
(<4), and then perform the rest of the copy a word at a time. From
looking at the generated assembly, this strategy seems likely to
outperform the previous algorithm even for small copies, as it's
identical for the first 3 bytes and then switches to using `lw/sw` pairs
(1/4 as many) at a cost of a single shrli and a few bytes.

Similarly, memset was taught to fill bytes one at a time until a word
boundary, and then pays a small cost (one zext, a lui/add pair, and one
mul to multiply the byte value by 0x1010101, extending it to 32 bits) to
fill the rest of the destination a word at a time. We use memset twice
in the hot path, both times with a number of bytes divisible by the word
size, to set approximately 350 bytes. Again, we were paying about a
thousand cyles in overhead with the previous byte-for-byte strategy.

Overall, these changes saved approximately one thousand cycles on the
hot path, bringing us down from ~3k cycles per `spi::start_transmit`
call to just about 1k cycles. Prior to the HAL update, the cycle count
was incredibly consistent, suggesting we'd effectively isolated the hot
path from external interference. Following that change, we see some
transmit calls that cost as few as ~800 cycles, but others that take as
long as ~1200 cycles: still, probably worth it to stop crashing
~randomly.