feat: introduce a fourth horsemem: `memcmp`
Adds the fourth (and final?) mem* function to our little RAM-resident
family, more for symmetry and as a weird flex than anything else.

Is this a good idea? Almost certainly not: I have not measured an
improvement in performance from including this function (even though it
seems to be used in the inner loop for embedded graphics' drawing
routines), but here it is.