refactor: introduce spi::Instance enum
Previously, the state of the SPI peripheral was spread across two
statics: QSPI and SPI_DMA_TRANSFER. Depending on which value was
populated, it was necessary to go "fetch" the opposite state and run an
operation on it. Crucially, though, only one state or the other could be
populated at a time.

So, this change models that as a two-state enum with an intrusive
Option variant. That last bit is not very rust-y of me, I know, but it
feels clearer to me than having a long chain of `.take().take()`s to
unwrap the type onion one layer at a time. Also, it makes implementing
`Instance::replace_with` possible, so that's something.